refactor(libs): split CustomSearch code into multiple files
